+++ Changelog.md
@@ -16,6 +16,7 @@ Released on ?.
 - Fixed downloading images from weather pages that need a login
 - Disable swiping in track details when hovering over the track's height
   profile
+- Fixed displaying time scale in track height profile
 
 ## Version 1.14.0 "New Zealand 2022/2023" edition